✨ (@hzcore/gesture-catcher) allow seamless preventDefault changes
Previously, any prop change would rebuild the sensors, which
could lead to unexpected behaviors if props changed during a gesture.
Now, a gesture may cause a change to `preventDefault`,
which will no longer cause new sensors to be built.

Other changes, like `passive`, or adding or removing sensors, will
still cause the sensors to be rebuilt, but we might be able to
mitigate some of these.