💥 (@hzcore/gesture-catcher) replace gestureRef prop with innerRef
Previously, a `gestureRef` prop could be passed to GestureCatcher
in lieu of using the `gestureRef` passed to the render prop.
This was meant to allow easier composition of behaviors on the same
ref, but dictated that the prop be a ref object (as opposed to
a callback).

Now, the concept of the `innerRef` represents that same
composability (i.e., you tell GestureCatcher what ref to use), but
you are always required to use the `gestureRef` passed to the render
prop. This allows GestureCatcher to bind to the ref when it is defined,
but also allows the composer to use an existing ref of any type
(object or callback) as the `innerRef`.

All of this is to avoid requiring things like 'wrapper' divs for
refs in behavioral compositions.