💥 (@hzdg/dom-utils) fix getContainingElement for all positions
The containing element can be different, depending on the position
of the element that is contained.

For example, `static`, `relative`, or `sticky` elements will treat
the nearest element that creates a block formatting context as
the containing element, while `absolute` elements will treat the
nearest element that has a non-static position as the containing
element. Lastly, `fixed` elements will typically treat the viewport
as the containing element.